Adding refreshToken hook for showing tooltips when ALT is held
In the past this module used the 'hoverToken' hook to show tooltips when alt was held down because when foundry would highlight tokens while alt was held, it considered those tokens as hovered. But it seems at some point in v10, either the hook 'refreshToken' was added or started to be used when alt was held down to highlight tokens, causing tooltips to not be shown when alt was held down because the hoverToken handler was never being called in the first place. This fix simply adds the hoverToken handler to the refreshToken hook as well.